Move opening search exact endpoint to SearchEndpoint
Refactored the exact search for openings from OpeningEndpoint to a new SearchEndpoint at /api/search/openings. Updated related integration tests and improved schema annotations for OpeningSearchResponseDto. This change improves endpoint organization and test coverage.
