+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/entity/SilvicultureSearchProjection.java
@@ -39,7 +39,9 @@ public interface SilvicultureSearchProjection {
   String getClientLocation();
 
   LocalDate getRegenDelayDate();
+
   LocalDate getEarlyFreeGrowingDate();
+
   LocalDate getLateFreeGrowingDate();
 
   LocalDateTime getUpdateTimestamp();
