+++ backend/src/test/java/ca/bc/gov/restapi/results/oracle/endpoint/OpeningEndpointIntegrationTest.java
@@ -19,7 +19,6 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import org.hamcrest.Matchers;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -1115,218 +1114,5 @@ class OpeningEndpointIntegrationTest extends AbstractTestContainerIntegrationTes
             get("/api/openings/1589595/attachments/" + nonExistentGuid)
                 .accept(MediaType.APPLICATION_JSON))
         .andExpect(status().isNotFound());
-  }
-
-  @Test
-  @DisplayName("Opening search exact with sub opening number should succeed")
-  void openingSearchExact_subOpeningNumber_shouldSucceed() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?subOpeningNumber=514")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andExpect(jsonPath("$.page.number").value("0"))
-        .andExpect(jsonPath("$.page.size").value("20"))
-        .andExpect(jsonPath("$.page.totalElements").isNumber())
-        .andExpect(jsonPath("$.content").isArray())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with mapsheet grid should succeed")
-  void openingSearchExact_mapsheetGrid_shouldSucceed() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?mapsheetGrid=93")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andExpect(jsonPath("$.page.number").value("0"))
-        .andExpect(jsonPath("$.page.size").value("20"))
-        .andExpect(jsonPath("$.page.totalElements").isNumber())
-        .andExpect(jsonPath("$.content").isArray())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with opening ID should succeed")
-  void openingSearchExact_openingId_shouldSucceed() throws Exception {
-    String clientNumber = "00010003";
-    clientApiStub.stubFor(
-        WireMock.get(urlPathEqualTo("/clients/findByClientNumber/" + clientNumber))
-            .willReturn(
-                okJson(
-                    """
-                {
-                  "clientNumber": "00010003",
-                  "clientName": "MINISTRY OF FORESTS",
-                  "legalFirstName": null,
-                  "legalMiddleName": null,
-                  "clientStatusCode": "ACT",
-                  "clientTypeCode": "F",
-                  "acronym": "MOF"
-                }
-                """)));
-
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?openingId=101017")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andExpect(jsonPath("$.page.number").value("0"))
-        .andExpect(jsonPath("$.page.size").value("20"))
-        .andExpect(jsonPath("$.page.totalElements").isNumber())
-        .andExpect(jsonPath("$.content[0].openingId").value(101017L))
-        .andExpect(jsonPath("$.content[0].openingNumber").exists())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with no results should return empty")
-  void openingSearchExact_noResults_shouldReturnEmpty() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?openingId=999999999")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andExpect(jsonPath("$.page.number").value("0"))
-        .andExpect(jsonPath("$.page.size").value("20"))
-        .andExpect(jsonPath("$.page.totalElements").value("0"))
-        .andExpect(jsonPath("$.content", Matchers.empty()))
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with invalid mapsheet grid should return 400")
-  void openingSearchExact_invalidMapsheetGrid_shouldReturn400() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?mapsheetGrid=99")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isBadRequest())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with invalid mapsheet letter should return 400")
-  void openingSearchExact_invalidMapsheetLetter_shouldReturn400() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?mapsheetLetter=X")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isBadRequest())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with invalid mapsheet quad should return 400")
-  void openingSearchExact_invalidMapsheetQuad_shouldReturn400() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?mapsheetQuad=5")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isBadRequest())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with invalid mapsheet sub quad should return 400")
-  void openingSearchExact_invalidMapsheetSubQuad_shouldReturn400() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?mapsheetSubQuad=6")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isBadRequest())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with invalid entry date range should return 400")
-  void openingSearchExact_invalidEntryDateRange_shouldReturn400() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?entryDateStart=2025-01-16&entryDateEnd=2025-01-01")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isBadRequest())
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with valid entry date range should succeed")
-  void openingSearchExact_validEntryDateRange_shouldSucceed() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?entryDateStart=2001-01-01&entryDateEnd=2025-12-31")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andExpect(jsonPath("$.page.number").value("0"))
-        .andExpect(jsonPath("$.page.size").value("20"))
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with page size should succeed")
-  void openingSearchExact_pageSize_shouldSucceed() throws Exception {
-    mockMvc
-        .perform(
-            get("/api/openings/search-exact?openingId=101017&page=0&size=10")
-                .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType("application/json"))
-        .andReturn();
-  }
-
-  @Test
-  @DisplayName("Opening search exact with composed opening number should be formatted correctly")
-  void openingSearchExact_composedOpeningNumber_shouldBeFormatted() throws Exception {
-    String clientNumber = "00010003";
-    clientApiStub.stubFor(
-        WireMock.get(urlPathEqualTo("/clients/findByClientNumber/" + clientNumber))
-            .willReturn(
-                okJson(
-                    """
-                {
-                  "clientNumber": "00010003",
-                  "clientName": "MINISTRY OF FORESTS",
-                  "legalFirstName": null,
-                  "legalMiddleName": null,
-                  "clientStatusCode": "ACT",
-                  "clientTypeCode": "F",
-                  "acronym": "MOF"
-                }
-                """)));
-
-    MvcResult result =
-        mockMvc
-            .perform(
-                get("/api/openings/search-exact?openingId=101017")
-                    .header("Content-Type", MediaType.APPLICATION_JSON_VALUE)
-                    .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isOk())
-            .andExpect(content().contentType("application/json"))
-            .andReturn();
-
-    // Extract and verify the opening number is properly composed (not just the raw DB value)
-    String responseBody = result.getResponse().getContentAsString();
-    String openingNumber = JsonPath.read(responseBody, "$.content[0].openingNumber");
-    Assertions.assertNotNull(openingNumber);
-    // Should be like "92K 014 0.0  514" not just "514"
-    Assertions.assertTrue(
-        openingNumber.contains(" "), "Opening number should be composed with spaces");
   }
 }
